+++ src/pages/GeysersPage/components/GeyserListItem.tsx
@@ -2,13 +2,9 @@ import * as React from "react";
 import { GeyserType } from "oni-save-parser";
 import classnames from "classnames";
 
-import {
-  Theme,
-  createStyles,
-  withStyles,
-  WithStyles
-} from "@material-ui/core/styles";
+import { WithTranslation, withTranslation } from "react-i18next";
 
+import { Theme, makeStyles } from "@material-ui/core/styles";
 import Paper from "@material-ui/core/Paper";
 import Divider from "@material-ui/core/Divider";
 import Typography from "@material-ui/core/Typography";
@@ -25,69 +21,144 @@ export interface GeyserListItemProps {
   className?: string;
 }
 
-const styles = (theme: Theme) =>
-  createStyles({
-    root: {
-      width: theme.spacing(45),
-      display: "flex",
-      flexDirection: "column",
-      padding: theme.spacing(2)
-    },
-    titleBar: {
-      display: "flex",
-      flexDirection: "row",
-      marginBottom: theme.spacing()
-    },
-    titleControls: {
-      display: "flex",
-      flexDirection: "row",
-      marginLeft: "auto"
-    },
-    sliderSection: {
-      marginTop: theme.spacing()
-    },
-    valueLabel: {
-      marginBottom: theme.spacing()
-    }
-  });
+const useStyles = makeStyles((theme: Theme) => ({
+  root: {
+    width: theme.spacing(45),
+    display: "flex",
+    flexDirection: "column",
+    padding: theme.spacing(2)
+  },
+  titleBar: {
+    display: "flex",
+    flexDirection: "row",
+    marginBottom: theme.spacing()
+  },
+  titleControls: {
+    display: "flex",
+    flexDirection: "row",
+    marginLeft: "auto"
+  },
+  sliderSection: {
+    marginTop: theme.spacing()
+  },
+  valueLabel: {
+    marginBottom: theme.spacing()
+  }
+}));
 
-type Props = GeyserListItemProps & WithStyles<typeof styles>;
+type Props = GeyserListItemProps & WithTranslation;
 
-const GeyserListItem: React.FC<Props> = ({
-  classes,
-  className,
-  gameObjectId
-}) => {
-  const { geyserType, emitRate, onChangeEmitRate, onChangeGeyserType } = useGeyser(gameObjectId);
+const GeyserListItem: React.FC<Props> = ({ className, gameObjectId, t }) => {
+  const styles = useStyles();
+
+  const {
+    geyserType,
+    emitRate,
+    yearLength,
+    yearActive,
+    emitActive,
+    onChangeEmitRate,
+    onChangeGeyserType,
+    onChangeYearLength,
+    onChangeYearActive,
+    onChangeEmitActive
+  } = useGeyser(gameObjectId);
+
+  const onGeyserTypeSelected = React.useCallback(
+    (e: React.ChangeEvent<{ value: unknown }>) => {
+      onChangeGeyserType(e.target.value as string);
+    },
+    [onChangeGeyserType]
+  );
+
+  const onGeyserRateChanged = React.useCallback(
+    (_: any, value: unknown) => {
+      const emitRate = (value as number) / 100;
+      onChangeEmitRate(emitRate);
+    },
+    [onChangeEmitRate]
+  );
+
+  const onGeyserYearLengthChanged = React.useCallback(
+    (_: any, value: unknown) => {
+      const fraction = (value as number) / 100;
+      onChangeYearLength(fraction);
+    },
+    [onChangeYearLength]
+  );
+
+  const onGeyserYearActiveChanged = React.useCallback(
+    (_: any, value: unknown) => {
+      const fraction = (value as number) / 100;
+      onChangeYearActive(fraction);
+    },
+    [onChangeYearActive]
+  );
+
+  const onGeyserEmissionLengthChanged = React.useCallback(
+    (_: any, value: unknown) => {
+      const fraction = (value as number) / 100;
+      onChangeEmitActive(fraction);
+    },
+    [onChangeEmitActive]
+  );
+
   return (
-    <Paper className={classnames(className, classes.root)}>
-      <div className={classes.titleBar}>
+    <Paper className={classnames(className, styles.root)}>
+      <div className={styles.titleBar}>
         <Typography variant="h6">{geyserType}</Typography>
-        <div className={classes.titleControls} />
+        <div className={styles.titleControls} />
       </div>
       <Divider />
-      <Select
-        value={geyserType || ""}
-        onChange={e => onChangeGeyserType(e.target.value as string)}
-      >
+      <Select value={geyserType || ""} onChange={onGeyserTypeSelected}>
         {keysOfType(GeyserType).map(typeName => (
           <MenuItem key={typeName} value={typeName}>
             {typeName}
           </MenuItem>
         ))}
       </Select>
-      <div className={classes.sliderSection}>
-        <Typography className={classes.valueLabel} id="rate-label">
-          Rate
-          </Typography>
+      <div className={styles.sliderSection}>
+        <Typography className={styles.valueLabel} id="lifecycle-label">
+          {t("geyser.total_lifecycle_time")}
+        </Typography>
+        <Slider
+          aria-labelledby="lifecycle-label"
+          defaultValue={(yearLength || 0) * 100}
+          onChangeCommitted={onGeyserYearLengthChanged}
+        />
+      </div>
+      <div className={styles.sliderSection}>
+        <Typography className={styles.valueLabel} id="active-label">
+          {t("geyser.active_time")}
+        </Typography>
+        <Slider
+          aria-labelledby="active-label"
+          defaultValue={(yearActive || 0) * 100}
+          onChangeCommitted={onGeyserYearActiveChanged}
+        />
+      </div>
+      <div className={styles.sliderSection}>
+        <Typography className={styles.valueLabel} id="emitting-label">
+          {t("geyser.emission_length")}
+        </Typography>
+        <Slider
+          aria-labelledby="emitting-label"
+          defaultValue={(emitActive || 0) * 100}
+          onChangeCommitted={onGeyserEmissionLengthChanged}
+        />
+      </div>
+      <div className={styles.sliderSection}>
+        <Typography className={styles.valueLabel} id="rate-label">
+          {t("geyser.emission_rate")}
+        </Typography>
         <Slider
           aria-labelledby="rate-label"
           defaultValue={(emitRate || 0) * 100}
-          onChangeCommitted={(_, value) => onChangeEmitRate((value as number) / 100)}
+          onChangeCommitted={onGeyserRateChanged}
         />
       </div>
     </Paper>
   );
-}
+};
 
-export default withStyles(styles)(GeyserListItem);
+export default withTranslation()(GeyserListItem);