+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# vNext
+
+- Allow changing geyser year, active, and emission length.
+- Make geyser text translatable.
+- Add ability to delete individual loose materials.
+- Add ability to search material list.
+
 # 3.13.1
 
 - Update oni-save-parser: Open up restrictions on .NET name validator to support mods with symbols or unicode in property names.